refactored BadgeService deleted Interface
