+++ src/main/java/com/greenfox/lvlup/controller/AdminController.java
@@ -3,8 +3,6 @@ package com.greenfox.lvlup.controller;
 import com.greenfox.lvlup.exception.GeneralException;
 import com.greenfox.lvlup.exception.SuccessfulQuery;
 import com.greenfox.lvlup.model.entityTestingDto.BadgeDTO;
-import com.greenfox.lvlup.model.entity.Badge;
-import com.greenfox.lvlup.repositrory.BadgeRepository;
 import com.greenfox.lvlup.service.BadgeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -20,12 +18,10 @@ import javax.validation.Valid;
 @RestController
 public class AdminController {
     private BadgeService badgeService;
-    private BadgeRepository badgeRepository;
 
     @Autowired
-    public AdminController(BadgeService badgeService, BadgeRepository badgeRepository) {
+    public AdminController(BadgeService badgeService) {
         this.badgeService = badgeService;
-        this.badgeRepository = badgeRepository;
     }
 
     @PostMapping(value = "/admin/add",
@@ -36,14 +32,8 @@ public class AdminController {
         if (token == null || token.equals("")) {
             throw new GeneralException("Unauthorized", HttpStatus.UNAUTHORIZED);
         } else {
-            Badge badgeToCreate = badgeService.getBadgefromDTO(dtoToAdd);
-            System.out.println(badgeToCreate.getName());
-            System.out.println(badgeToCreate.getId());
-            System.out.println(badgeToCreate.getVersion());
-            System.out.println(badgeToCreate.getTag());
-            System.out.println(badgeToCreate.getLevels());
-            badgeService.createBadge(badgeToCreate);
-            //badgeRepository.save(badgeToCreate);
+            //Badge badgeToCreate = badgeService.getBadgefromDTO(dtoToAdd);
+            //badgeService.createBadge(badgeToCreate);
             return new ResponseEntity<>(new SuccessfulQuery("Success"), HttpStatus.CREATED);
         }
